requirements/common.txt
Django~=4.2.6
djangorestframework~=3.14.0
django-cors-headers~=4.3.0
drf-yasg~=1.21.7
python-dotenv~=1.0.0
Pillow~=10.3.0
requests~=2.31.0
boto3~=1.34.69
django-storages~=1.14.2
whitenoise~=6.6.0
psycopg2-binary~=2.9.9
celery~=5.4.0
redis~=5.0.4
